# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/users/addresses": {
            "get": {
                "description": "Post Addresses for users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Address"
                ],
                "summary": "Get Address",
                "parameters": [
                    {
                        "type": "string",
                        "format": "\"Bearer your_actual_token_here\"",
                        "description": "authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "description": "Post Addresses for users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Address"
                ],
                "summary": "Post Address",
                "parameters": [
                    {
                        "type": "string",
                        "format": "\"Bearer your_actual_token_here\"",
                        "description": "authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Address details",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.Addr"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/users/addresses/{id}": {
            "delete": {
                "description": "Delete a specific address for a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Address"
                ],
                "summary": "Delete Address by ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "\"hex\"",
                        "description": "Address ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "\"Bearer your_actual_token_here\"",
                        "description": "authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
        "/api/users/carts": {
            "get": {
                "description": "Get a product From cart",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "Get Cart",
                "parameters": [
                    {
                        "type": "string",
                        "format": "\"Bearer your_actual_token_here\"",
                        "description": "authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "description": "Post a product to cart",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "Post Cart",
                "parameters": [
                    {
                        "type": "string",
                        "format": "\"Bearer your_actual_token_here\"",
                        "description": "authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Comment details",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/response.Input"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success"
                    }
                }
            },
            "delete": {
                "description": "Get a product From cart",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "Get Cart",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ProductId",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "\"Bearer your_actual_token_here\"",
                        "description": "authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "entities.Addr": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "address": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "postalCode": {},
                "state": {
                    "type": "string"
                }
            }
        },
...
        "response.Input": {
            "type": "object",
            "properties": {
                "productId": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "quantityState": {
                    "type": "string"
                },
                "variationsKey": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
    }
}